requirements.txt
absl-py==0.12.0
argon2-cffi==20.1.0
asgiref==3.3.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.22
decorator==5.0.3
defusedxml==0.7.1
Django==3.1.7
docutils==0.16
entrypoints==0.3
flatbuffers==1.12
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mpmath==1.2.1
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nose==1.3.7
notebook==6.3.0
numpy==1.19.5
oauthlib==3.1.0
#object-detection==0.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.0
pkginfo==1.7.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
#pywin32==300
#pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
readme-renderer==29.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
scipy==1.6.2
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
sympy==1.7.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
tensorflow-hub==0.11.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2021032501
#tf-models-official==2.4.0
tf-slim==1.1.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
twine==3.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
gunicorn==19.9.0
